+++ packages/core/tests/content-smuggling.test.ts
@@ -6,6 +6,18 @@ import { describe, expect, it } from 'vitest'
 
 import { DiagnosticDetail } from '../src/diagnostics.js'
 import { ObservationEnvelope } from '../src/observations.js'
+import { OBSERVATION_SCHEMA_VERSION } from '../src/schema.js'
+import {
+  collectSessionMeta,
+  collectToolResultMeta,
+  compactEntry,
+  dedupeStreamingMessageIds,
+  emptySessionMeta,
+  groupIntoTurns,
+  parseJsonlLine,
+  toObservations,
+} from '../src/providers/claude/index.js'
+import type { JournalEntry, ToolResultMeta } from '../src/providers/claude/index.js'
 
 const here = dirname(fileURLToPath(import.meta.url))
 const goldenEnvelope = JSON.parse(
@@ -101,6 +113,90 @@ describe('content-smuggling guardrail: accepted output is secret-free', () => {
     for (const secret of ALL_SECRETS) {
       expect(serialized).not.toContain(secret)
     }
+  })
+})
+
+describe('content-smuggling guardrail: real claude decode -> toObservations is secret-free', () => {
+  // A hostile Claude transcript planting every secret in the free-text fields a
+  // real decode captures: the user prompt, a bash command, a Read file_path, the
+  // ai-title, the cwd, the git branch, and the project path. Decoding it fully
+  // and minimizing MUST surface none of them.
+  function decodeSession() {
+    const lines = [
+      JSON.stringify({
+        type: 'user',
+        timestamp: '2026-07-17T10:00:00.000Z',
+        sessionId: 'sess-hostile',
+        cwd: SECRETS.absPath,
+        gitBranch: 'feature/secret-plan',
+        // prompt, apiKey and file content all planted in the captured user text.
+        message: { role: 'user', content: `${SECRETS.prompt} ${SECRETS.apiKey} ${SECRETS.fileContent}` },
+      }),
+      JSON.stringify({
+        type: 'assistant',
+        timestamp: '2026-07-17T10:00:05.000Z',
+        sessionId: 'sess-hostile',
+        gitBranch: 'feature/secret-plan',
+        message: {
+          id: 'msg-hostile-1',
+          type: 'message',
+          role: 'assistant',
+          model: 'claude-opus-4-8',
+          usage: { input_tokens: 1200, output_tokens: 340, cache_read_input_tokens: 800, cache_creation_input_tokens: 120 },
+          content: [
+            { type: 'tool_use', id: 'tu1', name: 'Bash', input: { command: SECRETS.commandLine } },
+            { type: 'tool_use', id: 'tu2', name: 'Read', input: { file_path: SECRETS.absPath } },
+            // A hostile tool NAME carrying a command line (spaces + slashes): it
+            // fails the canonical charset and must be dropped, not emitted.
+            { type: 'tool_use', id: 'tu3', name: SECRETS.commandLine, input: {} },
+          ],
+        },
+      }),
+      JSON.stringify({ type: 'ai-title', sessionId: 'sess-hostile', aiTitle: SECRETS.prompt }),
+    ]
+
+    const raw = lines.map(l => parseJsonlLine(l)).filter((e): e is JournalEntry => e !== null)
+    const meta = emptySessionMeta()
+    const toolResultMeta = new Map<string, ToolResultMeta>()
+    for (const entry of raw) {
+      collectToolResultMeta(entry, toolResultMeta)
+      collectSessionMeta(entry, meta)
+    }
+    const compacted = raw.map(compactEntry)
+    const turns = groupIntoTurns(dedupeStreamingMessageIds(compacted), new Set<string>(), toolResultMeta)
+    return { turns, meta }
+  }
+
+  function buildEnvelope() {
+    const { turns, meta } = decodeSession()
+    const { sessions } = toObservations(
+      { sessionId: 'sess-hostile', projectPath: SECRETS.absPath, gitBranch: 'feature/secret-plan', isSidechain: meta.isSidechain, turns },
+      { privacyKey: 'test-privacy-key', provider: 'claude' },
+    )
+    return {
+      schemaVersion: OBSERVATION_SCHEMA_VERSION,
+      generator: { name: '@codeburn/core', version: '0.0.0-test' },
+      sessions,
+    }
+  }
+
+  it('produces a schema-valid envelope from the hostile transcript', () => {
+    expect(ObservationEnvelope.safeParse(buildEnvelope()).success).toBe(true)
+  })
+
+  it('the serialized envelope contains none of the planted secrets', () => {
+    const serialized = JSON.stringify(buildEnvelope())
+    for (const secret of ALL_SECRETS) {
+      expect(serialized).not.toContain(secret)
+    }
+  })
+
+  it('drops non-canonical (argument-carrying) tool names instead of emitting them', () => {
+    const env = buildEnvelope()
+    const allToolNames = env.sessions.flatMap(s => s.calls.flatMap(c => c.toolNames))
+    expect(allToolNames).toContain('Bash')
+    expect(allToolNames).toContain('Read')
+    expect(allToolNames).not.toContain(SECRETS.commandLine)
   })
 })
